+++ src/test/java/com/github/ddth/cql/qnd/QndHostDown.java
@@ -15,15 +15,31 @@ import com.github.ddth.cql.SessionManager;
  */
 public class QndHostDown {
 
+    static {
+        System.setProperty("org.slf4j.simpleLogger.logFile", "System.out");
+        System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "INFO");
+        System.setProperty("org.slf4j.simpleLogger.showThreadName", "false");
+        System.setProperty("org.slf4j.simpleLogger.showLogName", "false");
+        System.setProperty("org.slf4j.simpleLogger.showShortLogName", "false");
+    }
+
     public static void main(String[] args) throws Exception {
         try (SessionManager sm = new SessionManager()) {
+            sm.setDefaultHostsAndPorts("localhost").setDefaultUsername("test")
+                    .setDefaultPassword("test").setDefaultKeyspace(null);
             sm.init();
+
+            // initialize data
+            sm.executeNonSelect(
+                    "CREATE KEYSPACE IF NOT EXISTS test WITH REPLICATION={'class' : 'SimpleStrategy', 'replication_factor' : 1}");
+            sm.executeNonSelect("DROP TABLE IF EXISTS test.tbl_test");
+            sm.executeNonSelect("CREATE TABLE test.tbl_test (id text, name text, PRIMARY KEY(id))");
 
             long timestamp = 0;
             while (true) {
-                Session session = sm.getSession("localhost", "demo", "demo", "demo", false);
+                Session session = sm.getSession("localhost", "test", "test", null, false);
                 try {
-                    ResultSet rs = CqlUtils.execute(session, "SELECT * FROM tbldemo");
+                    ResultSet rs = CqlUtils.execute(session, "SELECT * FROM test.tbl_test");
                     Iterator<Row> it = rs.iterator();
                     while (it.hasNext()) {
                         Row row = it.next();
@@ -31,7 +47,7 @@ public class QndHostDown {
                     }
 
                     timestamp = System.currentTimeMillis();
-                    System.out.println("Ping: " + session);
+                    System.out.println("Session: " + session);
                     session.close();
                     Thread.sleep(2000);
                 } catch (Throwable e) {
@@ -40,6 +56,8 @@ public class QndHostDown {
                     // e.printStackTrace();
                     Thread.sleep(5000);
                 }
+
+                System.out.println();
             }
         }
     }