+++ src/test/java/com/github/ddth/cql/qnd/QndSessionManager.java
@@ -9,21 +9,21 @@ public class QndSessionManager {
         try (SessionManager sm = new SessionManager()) {
             sm.init();
 
-            Session session1 = sm.getSession("localhost", "demo", "demo", "demo", false);
+            Session session1 = sm.getSession("localhost", "test", "test", "test", false);
             System.out.println("Session 1 (forceNew = false / close = false): " + session1);
 
-            Session session2 = sm.getSession("localhost", "demo", "demo", "demo", false);
+            Session session2 = sm.getSession("localhost", "test", "test", "test", false);
             System.out.println("Session 2 (forceNew = false / close = false): " + session2);
         }
 
         try (SessionManager sm = new SessionManager()) {
             sm.init();
 
-            Session session1 = sm.getSession("localhost", "demo", "demo", "demo", false);
+            Session session1 = sm.getSession("localhost", "test", "test", "test", false);
             System.out.println("Session 1 (forceNew = false / close = true): " + session1);
             session1.close();
 
-            Session session2 = sm.getSession("localhost", "demo", "demo", "demo", false);
+            Session session2 = sm.getSession("localhost", "test", "test", "test", false);
             System.out.println("Session 2 (forceNew = false / close = true): " + session2);
             session2.close();
         }
@@ -31,10 +31,10 @@ public class QndSessionManager {
         try (SessionManager sm = new SessionManager()) {
             sm.init();
 
-            Session session1 = sm.getSession("localhost", "demo", "demo", "demo", true);
+            Session session1 = sm.getSession("localhost", "test", "test", "test", true);
             System.out.println("Session 1 (forceNew = true / close = false): " + session1);
 
-            Session session2 = sm.getSession("localhost", "demo", "demo", "demo", true);
+            Session session2 = sm.getSession("localhost", "test", "test", "test", true);
             System.out.println("Session 2 (forceNew = true / close = false): " + session2);
         }
     }